DOC reorder examples by file names
