+++ examples/treeple/treeple_tutorial_1_2_pvalue.py
@@ -1,7 +1,7 @@
 """
-================================
-1-2: Calculating p-value (MIGHT)
-================================
+===========================
+Calculating p-value (MIGHT)
+===========================
 """
 
 import matplotlib.pyplot as plt
